requirements.txt
fastapi
uvicorn
pydantic
langchain
langgraph
langchain-anthropic
python-dotenv
httpx
streamlit
ray
trafilatura
ddgs
sqlalchemy
